+++ CHANGES.md
@@ -1,7 +1,10 @@
 # Release History
 
-## 1.51.1 (2026-mm-dd)
+## 1.52.0 (2026-mm-dd)
 
+- **ENH: Add the possibility to read a feature layer from a GDB in `vectors`**
+- **ENH: Add the possibility to get the rasterio-compatible path from a raster layer stored in a GDB**
+- FIX: Correctly handle returned path from `arcpy.gp_layer_to_path` in case of a raster stored in a GDB
 - FIX: Force dask in `rasters.read`'s reprojection only if `chunks` is not None
 - FIX: Fix expand dim in case `xarray-spatial` is not usable
 - FIX: Don't force rasterio's presence in env in types.py